pyproject.toml
[project]
name = "v-router"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ben Selleslagh", email = "ben@vectrix.ai" }
]
requires-python = ">=3.13.3"
dependencies = [
    "anthropic[vertex]>=0.52.0",
    "colorlog>=6.9.0",
    "google-cloud-aiplatform>=1.94.0",
    "google-genai>=1.16.1",
    "openai>=1.82.0",
]

[project.scripts]
v-router = "v_router:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "bump2version>=1.0.1",
    "ipykernel>=6.29.5",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.11",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
    "D100",
    "D203",
    "D213",
]

exclude = ["examples/*", "tests/*", "scripts/*", "src/cloud-functions/*"]
